api/requirements.txt
django==5.2.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
celery==5.5.3
django-celery-beat==2.8.1
channels==4.2.2
uvicorn==0.34.2
websockets==15.0.1
channels_redis==4.2.1
django-celery-results==2.6.0
